refactor: 쿠폰 관련 vuex 로직 분리
