+++ src/store/index.js
@@ -1,16 +1,7 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
 import {account, setAuthInHeader} from '../api'
-// import {
-//     fetchProducts, fetchProduct, fetchReviews, fetchReview,
-//     fetchQnas, fetchQna, addReview, fetchOrderLog,
-//     addQuestion, fetchAccountCarts, deleteCartItem, addOrderItems,
-//     fetchProductsByKeyword, fetchProductsByCategory, fetchOrderInfo, addCupon, fetchCuponsByAccount,
-//     addAdress, fetchAdressByAccount, addPayOrder, fetchMyOrderLog, cancelOrder, fetchQuestionsByAccount,
-//     deleteQuestion, deleteAddress
-// } from '../api/index'
 import {
-    addCupon, fetchCuponsByAccount,
     addAdress, fetchAdressByAccount, deleteAddress
 } from '../api/index'
 import products from "./modules/products";
@@ -18,6 +9,7 @@ import reviews from "./modules/reviews";
 import questions from "./modules/questions";
 import orders from "./modules/orders";
 import carts from "./modules/carts";
+import cupons from "./modules/cupons";
 
 Vue.use(Vuex);
 
@@ -27,13 +19,13 @@ const store = new Vuex.Store({
         products,
         questions,
         orders,
-        carts
+        carts,
+        cupons
     },
     state: {
         account: {},
         accessToken: null,
         myAddresses: {},
-        myCupons: {}
     },
     mutations: {
         LOGIN(state, {accessToken}) {
@@ -48,32 +40,16 @@ const store = new Vuex.Store({
             delete localStorage.accessToken;
             setAuthInHeader(null);
         },
-        SET_CUPONS(state, cupons) {
-            state.orderInfo.accountCupons = cupons;
-        },
+
         SET_MY_ADDRESSES(state, myAddresses) {
             state.myAddresses = myAddresses;
         },
-        SET_MY_CUPONS(state, myCupons) {
-            state.myCupons = myCupons;
-        }
+
     },
     actions: {
         async LOGIN({commit}, payload) {
             const response = await account.login(payload);
             commit('LOGIN', response.data);
-            return response;
-        },
-        // async ADD_ORDER_ITEMS(_, payload) {
-        //     const response = await addOrderItems(payload);
-        //     return response;
-        // },
-        async ADD_CUPON(_, payload) {
-            const response = await addCupon(payload);
-            return response;
-        },
-        async FETCH_CUPONS_BY_ACCOUNT(payload) {
-            const response = await fetchCuponsByAccount(payload);
             return response;
         },
         async ADD_ADDRESS(_, payload) {
@@ -88,7 +64,6 @@ const store = new Vuex.Store({
             const response = await deleteAddress(payload);
             return response;
         }
-
     },
     getters: {
         isAuthenticated(state) {
